マップをホームに昇格・ナビからマップタブを削除
- / → /map へリダイレクト（マップが新しいホーム）
- 旧ランディングページの内容を /about に移動
- NavigationBar からマップ項目を削除（相談・お店を探す の2項目に）
